updated goals layout (see extended description)
when I tried to cross out the goal to reach out to a finance person turned dev, I could not get the layout to update correctly. I realized that I had started two ordered lists in a row, and that it was perhaps messing with the formatting and ability to cross out the goals. As such, I merged the ordered lists.